+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to
     description: 'Missing permission users.read',
   });
   ```
+- Added support for relative paths in `yarn j1-integration *` commands
 
 ## [7.0.0] - 2021-10-05